Correct ch4. ex4.(c)-(d) answers
the training Error is going to be lower for the more flexible model no
matter what the population regression line is (real f(x)). However the
test RSS will highly depend on the type of dataset underlying model and
answer to (d) is: not enough information is given b/c the form is
unknown.